requirements.txt
anyio==3.3.4
appdirs==1.4.3
apturl==0.5.2
asgiref==3.4.1
attrs==19.3.0
Automat==0.8.0
backcall==0.1.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
blinker==1.4
boto==2.49.0
Brlapi==0.7.0
cached-property==1.5.1
certifi==2019.11.28
chardet==3.0.4
click==8.0.3
colorama==0.4.3
command-not-found==0.3
constantly==15.1.0
cryptography==2.8
cssselect==1.1.0
cupshelpers==1.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
Django==2.2.12
dnspython==2.1.0
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.12.0
entrypoints==0.3
evdev==1.3.0
fastapi==0.70.0
fasteners==0.14.1
filelock==3.0.12
Flask==2.0.2
future==0.18.2
h11==0.12.0
html5lib==1.0.1
httplib2==0.14.0
httptools==0.2.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
iotop==0.6
ipython==7.13.0
ipython-genutils==0.2.0
itemadapter==0.2.0
itemloaders==1.0.4
itsdangerous==2.0.1
jedi==0.15.2
Jinja2==3.0.2
jmespath==0.10.0
joblib==1.0.1
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
llvmlite==0.37.0
lockfile==0.12.2
louis==3.12.0
lxml==4.5.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.0.1
monotonic==1.5
more-itertools==4.2.0
motor==2.5.1
mysqlclient==1.4.4
netifaces==0.10.4
numba==0.54.1
numpy==1.20.3
oauthlib==3.1.0
olefile==0.46
packaging==20.9
pandas==1.3.4
paramiko==2.6.0
parsel==1.5.2
parso==0.5.2
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
prompt-toolkit==2.0.10
protobuf==3.6.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyAudio==0.2.11
pycairo==1.16.2
pycups==1.9.73
pydantic==1.8.2
PyDispatcher==2.0.5
Pygments==2.3.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
pymongo==3.12.1
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.15.5
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-dotenv==0.19.1
python-magic==0.4.16
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
pyzmq==22.0.3
queuelib==1.5.0
redis==3.5.3
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
resampy==0.2.2
scipy==1.7.1
Scrapy==1.7.3
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
service-identity==18.1.0
setproctitle==1.1.10
simplejson==3.16.0
six==1.14.0
sniffio==1.2.0
soupsieve==1.9.5
sqlparse==0.2.4
starlette==0.16.0
systemd-python==234
termcolor==1.1.0
texttable==1.6.2
threadpoolctl==2.1.0
traitlets==4.3.3
Twisted==18.9.0
typing-extensions==3.7.4.3
ubuntu-advantage-tools==27.8
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
uvicorn==0.15.0
uvloop==0.16.0
vboxapi==1.0
virtualenv==20.0.17
w3lib==1.21.0
wadllib==1.3.3
watchgod==0.7
wcwidth==0.1.8
webencodings==0.5.1
websocket-client==0.53.0
websockets==10.0
Werkzeug==2.0.2
xkit==0.0.0
zipp==1.0.0
zope.interface==4.7.1
